+++ .idea/modules.xml
@@ -3,6 +3,8 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/.idea/java-calculator-1.iml" filepath="$PROJECT_DIR$/.idea/java-calculator-1.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/java-calculator-1.main.iml" filepath="$PROJECT_DIR$/.idea/modules/java-calculator-1.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/java-calculator-1.test.iml" filepath="$PROJECT_DIR$/.idea/modules/java-calculator-1.test.iml" />
     </modules>
   </component>
 </project>
