--- Operators.java
@@ -1,7 +1,0 @@
-public class Operators {
-    Long operator;
-
-    public Operators(Long operator){
-        this.operator = operator;
-    }
-}
