--- View/OutputView.java
@@ -1,4 +1,0 @@
-package View;
-
-public class OutputView {
-}
