+++ src/main/java/org/example/Operators.java
@@ -1,0 +1,9 @@
+package org.example;
+
+public class Operators {
+    Long operator;
+
+    public Operators(Long operator){
+        this.operator = operator;
+    }
+}
