+++ src/main/java/org/example/View/OutputView.java
@@ -1,0 +1,4 @@
+package org.example.View;
+
+public class OutputView {
+}
